requirements.txt
fastapi==0.111.1
uvicorn==0.15.0
sqlalchemy==1.4.23
pydantic>=2.0.1
pydantic-settings==2.0.3
alembic==1.7.1
psycopg2-binary==2.9.1
python-jose==3.3.0
passlib==1.7.4
python-multipart>=0.0.7
email-validator>=2.0.0
transformers==4.30.2
torch==1.13.1